fix sponsor card hover problem
the usual styles for links in content sections (section .general-info a)
was being applied to these sponsor cards on a content page, not an issue
for spnsors page. I couldn't find a good way to override the above rule
or rewrite it so I just used !important—I'm open to better solutions.